+++ src/pages/home.tsx
@@ -5,7 +5,7 @@ import useIsLoggedIn from '@/hooks/use-is-logged-in'
 import { useNavigate, useSearchParams } from 'react-router-dom'
 import { useEffect, useState } from 'react'
 import Challenges from '@/components/home-screen/challenges'
-import PageContainer from '@/components/common/page-container'
+import PageContainer from '@/components/common/page-layout/container'
 import { Button } from '@/components/ui/button'
 import WarnNotLoggedIn from '@/components/home-screen/warn-not-logged-in'
 import PageHeaderSection from '@/components/common/page-header-section'
