+++ src/pages/my-page/my-page-layout.tsx
@@ -1,4 +1,4 @@
-import PageContainer from '@/components/common/page-container'
+import PageContainer from '@/components/common/page-layout/container'
 import PageHeaderSection from '@/components/common/page-header-section'
 import PageTitle from '@/components/common/page-title'
 import { ReactNode } from 'react'
